refactor: 도메인 메서드 객체로 변경
추후 도메인 로직의 확장을 고려하여 객체로 변경
